ClimbingPathBot/requirements.txt
openai==1.52.2
python-telegram-bot==21.6